+++ src/App.vue
@@ -1,33 +1,33 @@
 <template>
-	<div id="app">
-		<transition name="fade">
-			<router-view />
-		</transition>
-	</div>
+  <div id="app">
+    <transition name="fade">
+      <router-view />
+    </transition>
+  </div>
 </template>
 
 <style>
 #app {
-	font-family: Avenir, Helvetica, Arial, sans-serif;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-	color: #2c3e50;
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  color: #2c3e50;
 }
 
 #nav a {
-	font-weight: bold;
-	color: #2c3e50;
+  font-weight: bold;
+  color: #2c3e50;
 }
 
 #nav a.router-link-exact-active {
-	color: #42b983;
+  color: #42b983;
 }
 
 .fade-enter-active {
-	transition: opacity 0.7s;
+  transition: opacity 0.7s;
 }
 
 .fade-enter {
-	opacity: 0;
+  opacity: 0;
 }
 </style>
